backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.10.4
openai==1.58.1
httpx==0.28.1
python-dotenv==1.0.1
sse-starlette==2.2.1
stripe==11.4.1
